Single pass scope validator
