+++ Code/Language/m/src/m/validation/Validator.java
@@ -5,41 +5,37 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import m.validation.Problem.Severity;
+
 import m.library.Library;
 import m.model.Game;
 
 public class Validator
 {
-	Result result;
-
 	Parser parser;
 	ScopeValidator scopeValidator;
 	TypeValidator typeValidator;
 
 	public Validator()
 	{
-		var library = Library.ENGLISH;
-
-		result = new Result(new Game(library), new HashMap<>());
 		parser = new Parser();
-		scopeValidator = new ScopeValidator(library);
+		scopeValidator = new ScopeValidator();
 		typeValidator = new TypeValidator();
 	}
 
 	public Result validate(String file, String text)
 	{
+		var result = new Result(new Game(Library.ENGLISH), new HashMap<>());
+
 		var parsed = parser.validate(text);
 
 		var scoped = scopeValidator.validate(file, parsed.file);
 
 		var typed = typeValidator.validate(file, parsed.file, scoped.expressionGraph);
 
-		if (!parsed.problems.isEmpty())
-		{
-			result.problems.put(file, parsed.problems);
-		}
-		scoped.problems.forEach((f,problems)-> result.problems.getOrDefault(f, new ArrayList<>()).addAll(problems));
-		typed .problems.forEach((f,problems)-> result.problems.getOrDefault(f, new ArrayList<>()).addAll(problems));
+		result.problems.put(file, parsed.problems);
+
+		merge(result.problems, scoped.problems, typed.problems);
 
 		result.game.components.putAll(typed.components);
 		result.game.functions.addAll(typed.functions);
@@ -49,18 +45,53 @@ public class Validator
 
 	public Result delete(String file)
 	{
-		return result;
+		var result = new Result(new Game(Library.ENGLISH), new HashMap<>());
+
+		var scoped = scopeValidator.delete(file);
+
+		var typed = typeValidator.delete(file);
+
+		merge(result.problems, scoped.problems, typed.problems);
+
+		result.game.components.putAll(typed.components);
+		result.game.functions.addAll(typed.functions);
+
+		return new Result(null,null);
+	}
+
+	private void merge(Map<String,List<Problem>> result, Map<String,List<Problem>> scope, Map<String,List<Problem>> types)
+	{
+		scope.forEach((f,problems)->
+		{
+			result.computeIfAbsent(f, k->new ArrayList<>()).addAll(problems);
+		});
+		types.forEach((f,problems)->
+		{
+			result.computeIfAbsent(f, k -> new ArrayList<>()).addAll(problems);
+		});
 	}
 
 	public static class Result
 	{
 		public Game game;
 		public Map<String,List<Problem>> problems;
+		public boolean hasErrors;
 
 		public Result(Game game, Map<String,List<Problem>> problems)
 		{
 			this.game = game;
 			this.problems = problems;
+
+			problems.forEach((file,list)->
+			{
+				for (var problem : list)
+				{
+					if (problem.severity == Severity.ERROR)
+					{
+						hasErrors = true;
+					}
+				}
+			});
 		}
 	}
 }
