+++ Documentation/2. Developer guide/4. Features.md
@@ -101,16 +101,13 @@ Known issues
 
 Privately
 
-- Detect if project has error diagnostics
 - Analyze result: generate project if no errors
-- Return list of diagnostics
 - Convert m.validation.problems to lsp4j.diagnostics
 - Convert nodes to text ranges
-- Convert offsets
+- Convert offsets to line + character text positions
 
 Known issues
 
-- ! Detecting if the globalData has errors can be done better with streams !
 - ! Converting nodes to text ranges can be done better by using NodeModel better !
 
 ## Validation package
@@ -119,12 +116,52 @@ Known issues
 
 ### Validator.java
 
-- Create and maintain previous calls memory, pass reference to filechecker and typechecker
-- Clear modified files every call
-- Locally validate modified file and add modified files
-- Globally check function names and add modified files
-- Globally check types and add modified files
-- Return reference to memory
+- Modify file: Validate syntax, scope and types
+- Delete file: Signal deletion to scope and types
+- Merge all problems in result.problems
+- Save typed functions and components in result.game
+- Detect if result has problems with error severity to signal validity of model
+
+### ScopeValidator.java
+
+- Clear obsolete memory
+- Walk the parsed model
+
+Variables
+
+- Declare variables: Query entity and assigning value for the first time
+- Standard variable: Redefinition error
+- Declared variable: Bind to declaration
+- Undeclared variable: Declare it
+- Access variable: In value expressions and cell entities
+- Standard variable : Type it
+- Declared variable: Bind to declaration
+- Undeclared variable: Undefined error
+
+Components
+
+- Declare component: Cells for the first time
+- Standard component: Do nothing
+- Declared component: Do nothing
+- Undeclared component: Declare it
+- Access component: Cells
+- Standard component: Type it
+- Declared component: Bind to declaration
+- ! Undeclared component: Never happens !
+
+Functions
+
+- ! Declare functions: Never happens !
+- Access function: In assignment, unary, binary and applications
+- Standard function with same number of arguments: Type it, bind same type variables
+- Non standard function: Undefined error in name
+
+Blocks
+
+- Standard block: Type it
+- Non standard block: Undefined error
+- Standard binding block: Type it
+- Non standard block: Undefined error
 
 ### TypeChecker.java
